Update template logic
To keep the template simple and not use many function tags we
pre-process the resources list on initialization to add the properties
that the template will need. Some of these rely on server side values,
so we compute these in the Jinja2 and pass them as parameters to the
module.